# fix: improve performance of ApplyPowerTableDiffs in ImportSnapshotToDatastore