The `sort` operation in `ApplyPowerTableDiffs` turns out to be expensive. 
This PR introduces another `func ApplyPowerTableDiffsToMap(powerTableMap map[gpbft.ActorID]gpbft.PowerEntry, diffs ...PowerTableDiff) (map[gpbft.ActorID]gpbft.PowerEntry, error)` to avoid sorting.

It also adds more validations against first instance, latest instance and powertable cid.

Perf stats of importing a mainnet F3 snapshot on my laptop show ~99% perf gain

```
# PR
0.884s
0.773s
0.864s

# main
86.547s
88.422s
86.998s
```